requirement.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.0
pdfplumber==0.10.2
requests==2.31.0
pydantic==2.4.0
python-dotenv==1.0.0